# Add Prepare API's to Rego objects