--- a/rego/rego_test.go
+++ b/rego/rego_test.go
@@ -13,11 +13,47 @@ import (
 
 	"github.com/open-policy-agent/opa/ast"
 	"github.com/open-policy-agent/opa/metrics"
+	"github.com/open-policy-agent/opa/storage"
 	"github.com/open-policy-agent/opa/storage/inmem"
 	"github.com/open-policy-agent/opa/topdown"
 	"github.com/open-policy-agent/opa/types"
 	"github.com/open-policy-agent/opa/util"
 )
+
+func assertEval(t *testing.T, r *Rego, expected string) {
+	t.Helper()
+	rs, err := r.Eval(context.Background())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+	assertResultSet(t, rs, expected)
+}
+
+func assertPreparedEvalQueryEval(t *testing.T, pq PreparedEvalQuery, options []EvalOption, expected string) {
+	t.Helper()
+	rs, err := pq.Eval(context.Background(), options...)
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+	assertResultSet(t, rs, expected)
+}
+
+func assertResultSet(t *testing.T, rs ResultSet, expected string) {
+	t.Helper()
+	result := []interface{}{}
+
+	for i := range rs {
+		values := []interface{}{}
+		for j := range rs[i].Expressions {
+			values = append(values, rs[i].Expressions[j].Value)
+		}
+		result = append(result, values)
+	}
+
+	if !reflect.DeepEqual(result, util.MustUnmarshalJSON([]byte(expected))) {
+		t.Fatalf("Expected:\n\n%v\n\nGot:\n\n%v", expected, result)
+	}
+}
 
 func TestRegoEvalExpressionValue(t *testing.T) {
 
@@ -98,34 +134,13 @@ func TestRegoEvalExpressionValue(t *testing.T) {
 		},
 	}
 
-	ctx := context.Background()
-
 	for _, tc := range tests {
 		t.Run(tc.query, func(t *testing.T) {
 			r := New(
 				Query(tc.query),
 				Module("", module),
 			)
-
-			rs, err := r.Eval(ctx)
-
-			if err != nil {
-				t.Fatalf("Unexpected error: %v", err)
-			}
-
-			result := []interface{}{}
-
-			for i := range rs {
-				values := []interface{}{}
-				for j := range rs[i].Expressions {
-					values = append(values, rs[i].Expressions[j].Value)
-				}
-				result = append(result, values)
-			}
-
-			if !reflect.DeepEqual(result, util.MustUnmarshalJSON([]byte(tc.expected))) {
-				t.Fatalf("Expected:\n\n%v\n\nGot:\n\n%v", tc.expected, result)
-			}
+			assertEval(t, r, tc.expected)
 		})
 	}
 }
@@ -155,7 +170,6 @@ func TestRegoInputs(t *testing.T) {
 		"nil":                {nil, `[[null]]`},
 		"slice of interface": {[]interface{}{"a", 2, true}, `[[["a", 2, true]]]`},
 	}
-	ctx := context.Background()
 
 	for desc, tc := range tests {
 		t.Run(desc, func(t *testing.T) {
@@ -163,23 +177,7 @@ func TestRegoInputs(t *testing.T) {
 				Query("input"),
 				Input(tc.input),
 			)
-			rs, err := r.Eval(ctx)
-			if err != nil {
-				t.Fatalf("Unexpected error: %v", err)
-			}
-
-			result := []interface{}{}
-			for i := range rs {
-				values := []interface{}{}
-				for j := range rs[i].Expressions {
-					values = append(values, rs[i].Expressions[j].Value)
-				}
-				result = append(result, values)
-			}
-
-			if !reflect.DeepEqual(result, util.MustUnmarshalJSON([]byte(tc.expected))) {
-				t.Fatalf("Expected:\n\n%v\n\nGot:\n\n%v", tc.expected, result)
-			}
+			assertEval(t, r, tc.expected)
 		})
 	}
 }
@@ -304,6 +302,339 @@ func TestPartialRewriteEquals(t *testing.T) {
 		t.Errorf("expected 1 query but found %d: %+v", len(pq.Queries), pq)
 	}
 	if pq.Queries[0].String() != expectedQuery {
-		t.Errorf("unexpected query in result, expected='%s' found='%s'", expectedQuery, pq.Queries[0].String())
+		t.Errorf("unexpected query in result, expected='%s' found='%s'",
+			expectedQuery, pq.Queries[0].String())
 	}
+}
+
+func TestPrepareAndEvalNewInput(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+	)
+
+	pq, err := r.PrepareForEval(context.Background())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+	}, "[[1]]")
+}
+
+func TestPrepareAndEvalNewMetrics(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	originalMetrics := metrics.New()
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+		Metrics(originalMetrics),
+	)
+
+	pq, err := r.PrepareForEval(context.Background())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	if len(originalMetrics.All()) == 0 {
+		t.Errorf("Expected metrics stored on 'originalMetrics' after Prepare()")
+	}
+
+	// Reset the original ones (for testing)
+	// and make a new one for the Eval
+	originalMetrics.Clear()
+	newMetrics := metrics.New()
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+		EvalMetrics(newMetrics),
+	}, "[[1]]")
+
+	if len(originalMetrics.All()) > 0 {
+		t.Errorf("Expected no metrics stored on original Rego object metrics but found: %s",
+			originalMetrics.All())
+	}
+
+	if len(newMetrics.All()) == 0 {
+		t.Errorf("Expected metrics stored on 'newMetrics' after Prepare()")
+	}
+}
+
+func TestPrepareAndEvalNewTransaction(t *testing.T) {
+	module := `
+	package test
+	x = data.foo.y
+	`
+	ctx := context.Background()
+	store := inmem.New()
+	txn := storage.NewTransactionOrDie(ctx, store, storage.WriteParams)
+
+	path, ok := storage.ParsePath("/foo")
+	if !ok {
+		t.Fatalf("Unexpected error parsing path")
+	}
+
+	err := storage.MakeDir(ctx, store, txn, path)
+	if err != nil {
+		t.Fatalf("Unexpected error writing to store: %s", err.Error())
+	}
+
+	err = store.Write(ctx, txn, storage.AddOp, path, map[string]interface{}{"y": 1})
+	if err != nil {
+		t.Fatalf("Unexpected error writing to store: %s", err.Error())
+	}
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Store(store),
+		Transaction(txn),
+	)
+
+	pq, err := r.PrepareForEval(ctx)
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, nil, "[[1]]")
+	store.Commit(ctx, txn)
+
+	// Update the store directly and get a new transaction
+	newTxn := storage.NewTransactionOrDie(ctx, store, storage.WriteParams)
+	err = store.Write(ctx, newTxn, storage.ReplaceOp, path, map[string]interface{}{"y": 2})
+	if err != nil {
+		t.Fatalf("Unexpected error writing to store: %s", err.Error())
+	}
+	defer store.Abort(ctx, newTxn)
+
+	// Expect that the old transaction and new transaction give
+	// different results.
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{EvalTransaction(txn)}, "[[1]]")
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{EvalTransaction(newTxn)}, "[[2]]")
+}
+
+func TestPrepareAndEvalIdempotent(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+	)
+
+	pq, err := r.PrepareForEval(context.Background())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	// Expect evaluating the same thing >1 time gives the same
+	// results each time.
+	for i := 0; i < 5; i++ {
+		assertPreparedEvalQueryEval(t, pq, []EvalOption{
+			EvalInput(map[string]int{"y": 1}),
+		}, "[[1]]")
+	}
+}
+
+func TestPrepareAndEvalOriginal(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+		Input(map[string]int{"y": 2}),
+	)
+
+	pq, err := r.PrepareForEval(context.Background())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+	}, "[[1]]")
+
+	// Even after prepare and eval with different input
+	// expect that the original Rego object behaves
+	// as expected for Eval.
+
+	assertEval(t, r, "[[2]]")
+}
+
+func TestPrepareAndPartialResult(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+		Input(map[string]int{"y": 2}),
+	)
+
+	ctx := context.Background()
+
+	pq, err := r.PrepareForEval(ctx)
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+	}, "[[1]]")
+
+	// Even after prepare and eval with different input
+	// expect that the original Rego object behaves
+	// as expected for PartialResult.
+
+	partial, err := r.PartialResult(ctx)
+
+	r2 := partial.Rego(
+		Input(map[string]int{"y": 7}),
+	)
+	assertEval(t, r2, "[[7]]")
+}
+
+func TestPrepareWithPartialEval(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+	)
+
+	ctx := context.Background()
+
+	// Prepare the query and partially evaluate it
+	pq, err := r.PrepareForEval(ctx, WithPartialEval())
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+	}, "[[1]]")
+}
+
+func TestPrepareAndPartial(t *testing.T) {
+	mod := `
+	package test
+	default p = false
+	p {
+		input.x = 1
+	}
+	`
+	r := New(
+		Query("data.test.p == true"),
+		Module("test.rego", mod),
+	)
+
+	ctx := context.Background()
+
+	pq, err := r.PrepareForEval(ctx)
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"x": 1}),
+	}, "[[true]]")
+
+	// Even after prepare and eval with different input
+	// expect that the original Rego object behaves
+	// as expected for Partial.
+
+	partialQuery, err := r.Partial(ctx)
+	expectedQuery := "input.x = 1"
+	if len(partialQuery.Queries) != 1 {
+		t.Errorf("expected 1 query but found %d: %+v", len(partialQuery.Queries), pq)
+	}
+	if partialQuery.Queries[0].String() != expectedQuery {
+		t.Errorf("unexpected query in result, expected='%s' found='%s'",
+			expectedQuery, partialQuery.Queries[0].String())
+	}
+}
+
+func TestPrepareAndCompile(t *testing.T) {
+	module := `
+	package test
+	x = input.y
+	`
+
+	r := New(
+		Query("data.test.x"),
+		Module("", module),
+		Package("foo"),
+	)
+
+	ctx := context.Background()
+
+	pq, err := r.PrepareForEval(ctx)
+	if err != nil {
+		t.Fatalf("Unexpected error: %s", err.Error())
+	}
+
+	assertPreparedEvalQueryEval(t, pq, []EvalOption{
+		EvalInput(map[string]int{"y": 1}),
+	}, "[[1]]")
+
+	// Ensure that Compile still works after Prepare
+	// and its Eval has been called.
+	_, err = r.Compile(ctx)
+	if err != nil {
+		t.Errorf("Unexpected error when compiling: %s", err.Error())
+	}
+}
+
+func TestPartialResultWithInput(t *testing.T) {
+	mod := `
+	package test
+	default p = false
+	p {
+		input.x == 1
+	}
+	`
+	r := New(
+		Query("data.test.p"),
+		Module("test.rego", mod),
+	)
+
+	ctx := context.Background()
+	pr, err := r.PartialResult(ctx)
+
+	if err != nil {
+		t.Fatalf("unexpected error from Rego.PartialResult(): %s", err.Error())
+	}
+
+	r2 := pr.Rego(
+		Input(map[string]int{"x": 1}),
+	)
+
+	assertEval(t, r2, "[[true]]")
 }