There are a handful of commits changing things, each has some details in their commit messages. High level view of the changes:

* Add `PrepareForEval()` function to Rego objects. This will parse modules and queries, then compile the modules and query (in turn parsing the imports and packages, as needed). The results are stored on the Rego object which is referenced from a PartialResult for future use...
* Add `PrepareForPartial()` function to Rego objects. Similar to `PrepareForEval` it handles parsing/pre-compiling as much as possible, but tailored for doing `Partial()` evaluations.
*  Add `PreparedEvalQuery` object with an `Eval` function which can then use the prepared parts of the queries. This Eval differs from the one on the Rego object by (optionally) taking in a new input, metric, or transaction.
* Add `PreparedPartialQuery` object with a `Partial` function. Similar to `PreparedEvalQuery` but for `Partial`
* Decouple input and QueryCompiler, in turn decoupling the compiled query and input
* Remove the `QueryCompiler.WithInput()` method. As-is it isn't used for anything. There is a risk that clients embedding OPA and calling the compile API's directly are expecting input to be on the query context passed to custom stages registered in the compile step.
* Add new metric for input parsing, it is not longer counted as part of the query compilation step.


In addition there are some options to potentially expose this via the HTTP server. As-is only golang API users can get any benefits by preparing Rego objects and re-using the PartialResult.


I updated the authz benchmark to use this new API and the results are pretty good:

Before:
```
BenchmarkAuthzForbidAuthn-8      	    5000	    252037 ns/op	   87950 B/op	    1993 allocs/op
BenchmarkAuthzForbidPath-8       	    5000	    322503 ns/op	  113304 B/op	    2798 allocs/op
BenchmarkAuthzForbidMethod-8     	    5000	    329911 ns/op	  114517 B/op	    2833 allocs/op
BenchmarkAuthzAllow10Paths-8     	    5000	    326845 ns/op	  113506 B/op	    2813 allocs/op
BenchmarkAuthzAllow100Paths-8    	    2000	    816802 ns/op	  316721 B/op	    9688 allocs/op
BenchmarkAuthzAllow1000Paths-8   	     200	   6075679 ns/op	 2346239 B/op	   79022 allocs/op
```

After:
```
BenchmarkAuthzForbidAuthn-8      	   50000	     30648 ns/op	   10955 B/op	     224 allocs/op
BenchmarkAuthzForbidPath-8       	   20000	     97358 ns/op	   35683 B/op	    1029 allocs/op
BenchmarkAuthzForbidMethod-8     	   10000	    103638 ns/op	   36929 B/op	    1064 allocs/op
BenchmarkAuthzAllow10Paths-8     	   10000	    101128 ns/op	   35974 B/op	    1044 allocs/op
BenchmarkAuthzAllow100Paths-8    	    2000	    616450 ns/op	  239879 B/op	    7920 allocs/op
BenchmarkAuthzAllow1000Paths-8   	     300	   6076363 ns/op	 2269655 B/op	   77260 allocs/op
```


Fixes: #1305