@patrick-east 2019-04-17
Note that I still need to add new unit tests for rego.go. Updating the review to make sure we're still good on general direction.