@patrick-east 2019-04-23
Latest changes:
* Split Prepare into PrepareForEval and PrepareForPartial, each with specialized returned PreparedXXQuery objects.
* Refactor the normal Rego evaluation functions (except for Compile) to use Prepare and the resulting prepared query.