@patrick-east 2019-04-23
Rebased, updated changelog, and split/squashed the PR update commits.